Windows compilation fixes

+++ src/sputil/core/Field.cpp
@@ -71,7 +71,7 @@ void Field::setNull(VariantType vtype)
     if (vtype == VAR_NONE)
         m_dataType |= VAR_NULL;
     else
-        m_dataType = vtype | VAR_NULL;
+        m_dataType = uint16_t(vtype | VAR_NULL);
 }
 
 String Field::asString() const
